Add Jest setup and relevance tests
