+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "start": "ts-node gasguardian-userbot.ts",
     "build": "tsc",
-    "prisma": "prisma"
+    "prisma": "prisma",
+    "test": "jest"
   },
   "dependencies": {
     "dotenv": "^16.4.5",
@@ -21,6 +22,9 @@
   "devDependencies": {
     "typescript": "^5.4.5",
     "ts-node": "^10.9.2",
-    "prisma": "^6.8.2"
+    "prisma": "^6.8.2",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "@types/jest": "^29.5.3"
   }
 }
